refactor: replace keys with strings in section conf
